requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
google-generativeai==0.5.4
python-dotenv==0.21.1
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
prisma==0.11.0 # Specified version for Prisma Client Python
